+++ projects/ngx-fluent/src/public-api.ts
@@ -5,4 +5,5 @@
 export * from './lib/ngx-fluent.service';
 export * from './lib/ngx-fluent.pipe';
 export * from './lib/ngx-fluent.module';
+export * from './lib/ngx-fluent.provider';
 export * from './lib/types';